Remove automatic retries in case of errors
Currently this gems tries to do automatic retries to the next (or same)
endpoint in case of an error. This has two issues, stack overflow (retry
is by recursion) and not respecting timeouts. Stack overflow is
solvable, however the timeout issue not really given current
architecture.

Cleanest solution is to just rotate the endpoints and move the
responsibility for retry to the calling application.

Fixes davissp14/etcdv3-ruby#130
Fixes davissp14/etcdv3-ruby#131